+++ tests/Message/Request/StringableTest.php
@@ -32,7 +32,7 @@ class StringableTest extends TestCase
             ),
             Lines::ofContent('some body'),
         );
-        $stringable = new Stringable($request);
+        $stringable = Stringable::of($request);
 
         $this->assertInstanceOf(RequestInterface::class, $stringable);
         $this->assertSame($request->url(), $stringable->url());